+++ app/routes/__layout.tsx
@@ -1,4 +1,4 @@
-import type { ShouldReloadFunction } from "@remix-run/react";
+import type { ShouldRevalidateFunction } from "@remix-run/react";
 import { Outlet, useLoaderData } from "@remix-run/react";
 import type { LoaderFunction } from "@remix-run/server-runtime";
 import { json } from "@remix-run/server-runtime";
@@ -14,7 +14,7 @@ export const loader: LoaderFunction = async ({ request }) => {
   return json<LoaderData>({ allowSignUp });
 };
 
-export const unstable_shouldReload: ShouldReloadFunction = () => false;
+export const shouldRevalidate: ShouldRevalidateFunction = () => false;
 
 export default function Layout() {
   const data = useLoaderData<LoaderData>();