+++ app/routes/__public.tsx
@@ -1,4 +1,4 @@
-import type { ShouldReloadFunction } from "@remix-run/react";
+import type { ShouldRevalidateFunction } from "@remix-run/react";
 import { Outlet, useLoaderData } from "@remix-run/react";
 import type { LoaderFunction } from "@remix-run/server-runtime";
 import { json } from "@remix-run/server-runtime";
@@ -15,7 +15,7 @@ export const loader: LoaderFunction = async ({ request }) => {
   return json<LoaderData>({ allowSignUp });
 };
 
-export const unstable_shouldReload: ShouldReloadFunction = () => false;
+export const shouldRevalidate: ShouldRevalidateFunction = () => false;
 
 export default function PublicLayout() {
   const data = useLoaderData<LoaderData>();