+++ CHANGELOG.md
@@ -16,7 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- The main settings page now shows a list of the most commonly clicked on links
+- Move user information to a new stats page
+- The stats page now shows a list of the most commonly clicked on links
 
 ### Fixed